Trim action stack when neccesary. Fixes issue #6

+++ src/core/constants/actions.hpp
@@ -1,4 +1,4 @@
 #pragma once
 
-static constexpr size_t MAX_ACTION_STACK_SIZE = 200; // TODO: FIgure out if this is sufficient
+static constexpr unsigned int MAX_ACTION_STACK_SIZE = 200;
